+++ rollup.config.js
@@ -6,7 +6,7 @@ import typescript from "rollup-plugin-typescript2";
 const pkg = require("./package.json");
 
 export default {
-  input: `src/index.ts`,
+  input: "src/index.ts",
   output: [
     { file: pkg.main, name: "tsLuxon", format: "umd", sourcemap: true },
     { file: pkg.module, format: "cjs", sourcemap: true }