+++ src/types/public.ts
@@ -4,3 +4,7 @@ export * from "./info";
 export * from "./interval";
 export * from "./locale";
 export * from "./zone";
+
+import Intl from "./intl-2020";
+
+export { Intl };