+++ CHANGELOG.md
@@ -6,3 +6,8 @@
  
 ## 2.0.1
 * Fixed DateTime.hasSame comparison 
+
+## 2.0.2
+* Added tslint
+* Improved naming convention
+* Solved Intl issue when compiling Angular (improved Intl namespace declaring and merging until es2020.intl)